\name{JumpTest-package}
\alias{JumpTest-package}
\alias{JumpTest}
\docType{package}
\title{
\packageTitle{JumpTest}
}
\description{
\packageDescription{JumpTest}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{JumpTest}
\packageIndices{JumpTest}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{JumpTest}
Maintainer: \packageMaintainer{JumpTest}
}
\keyword{ nonparametric jump test, p-values pooling, false discovery rate control }