% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testsim.R
\name{jumptestday}
\alias{jumptestday}
\title{Nonparametric jump test for each interval}
\usage{
jumptestday(ret, method = "BNS")
}
\arguments{
\item{ret}{log return vector}

\item{method}{jump test methods, chosen from "BNS", "Amed", and "Amin"}
}
\value{
\item{stat}{test statistics}

\item{pvalue}{p-value}
}
\description{
perform nonparametric jump test for each given interval (day)
}
\examples{
orip <- runif(100)
testres <- jumptestday(orip)
ts <- testres@stat
pv <- testres@pvalue
}
