% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testsim.R
\name{pcombine}
\alias{pcombine}
\title{p-values matrix to be pooled}
\usage{
pcombine(retmat, method)
}
\arguments{
\item{retmat}{log return matrix by columns}

\item{method}{jump test methods, chosen from "BNS", "Amed", and "Amin"}
}
\value{
a p-values matrix
}
\description{
generate p-value matrix with given methods (at least 2)
}
\examples{
orip <- matrix(runif(3000),1000,3)
pmatrix <- pcombine(orip,c('BNS','Amed','Amin'))
}
