% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testsim.R
\name{SV}
\alias{SV}
\title{SV model with one factor simulation}
\usage{
SV(M, m, p0 = 3, mu = 0.05, v0 = 0, b = 0.2, alpha = 0.015,
  sigma = 0.05)
}
\arguments{
\item{M}{number of interverals to be simulated}

\item{m}{number of time points within each interval}

\item{p0}{start price}

\item{mu}{drift}

\item{v0}{starting volatility}

\item{b}{volatility parameter}

\item{alpha}{volatility parameter}

\item{sigma}{volatility parameter}
}
\value{
simulated time series
}
\description{
Simulate stochastic volatility model (np jump) with given length and other parameters
}
\examples{
SV(390,1200)
}
\references{
Yen, Y.-M. (2013). "Testing Jumps via False Discovery Rate Control." PloS one 8(4): e58365.
}
